Move VCF functions to io, add io_support for shell_command_runner
These VCF functions are specifically IO-related utilities, so moving them to a better scope.
