+++ augur/align.py
@@ -6,7 +6,8 @@ import os
 from shutil import copyfile
 import numpy as np
 from Bio import AlignIO, SeqIO, Seq, Align
-from .utils import run_shell_command, nthreads_value, shquote
+from .io import run_shell_command, shquote
+from .utils import nthreads_value
 from collections import defaultdict
 
 class AlignmentError(Exception):
