+++ augur/tree.py
@@ -15,8 +15,8 @@ import numpy as np
 from treetime.vcf_utils import read_vcf
 from pathlib import Path
 
-from .io import read_sequences
-from .utils import run_shell_command, nthreads_value, shquote, load_mask_sites
+from .io import read_sequences, run_shell_command, shquote
+from .utils import nthreads_value, load_mask_sites
 
 DEFAULT_ARGS = {
     "fasttree": "-nt -nosupport",
