+++ augur/utils.py
@@ -17,68 +17,12 @@ from augur.io import open_file
 from augur.util_support.color_parser import ColorParser
 from augur.util_support.metadata_file import MetadataFile
 from augur.util_support.node_data_reader import NodeDataReader
-from augur.util_support.shell_command_runner import ShellCommandRunner
 
 
 class AugurError(Exception):
     pass
 
 
-def is_vcf(filename):
-    """Convenience method to check if a file is a vcf file.
-
-    >>> is_vcf(None)
-    False
-    >>> is_vcf("./foo")
-    False
-    >>> is_vcf("./foo.vcf")
-    True
-    >>> is_vcf("./foo.vcf.GZ")
-    True
-    """
-    return bool(filename) and any(filename.lower().endswith(x) for x in ('.vcf', '.vcf.gz'))
-
-def read_vcf(filename):
-    if filename.lower().endswith(".gz"):
-        import gzip
-        file = gzip.open(filename, mode="rt", encoding='utf-8')
-    else:
-        file = open(filename, encoding='utf-8')
-
-    chrom_line = next(line for line in file if line.startswith("#C"))
-    file.close()
-    headers = chrom_line.strip().split("\t")
-    sequences = headers[headers.index("FORMAT") + 1:]
-
-    # because we need 'seqs to remove' for VCF
-    return sequences, sequences.copy()
-
-def write_vcf(input_filename, output_filename, dropped_samps):
-    if _filename_gz(input_filename):
-        input_arg = "--gzvcf"
-    else:
-        input_arg = "--vcf"
-
-    if _filename_gz(output_filename):
-        output_pipe = "| gzip -c"
-    else:
-        output_pipe = ""
-
-    drop_args = ["--remove-indv " + shquote(s) for s in dropped_samps]
-
-    call = ["vcftools"] + drop_args + [input_arg, shquote(input_filename), "--recode --stdout", output_pipe, ">", shquote(output_filename)]
-
-    print("Filtering samples using VCFTools with the call:")
-    print(" ".join(call))
-    run_shell_command(" ".join(call), raise_errors = True)
-    # remove vcftools log file
-    try:
-        os.remove('out.log')
-    except OSError:
-        pass
-
-def _filename_gz(filename):
-    return filename.lower().endswith(".gz")
 
 def get_json_name(args, default=None):
     if args.output_node_data:
@@ -301,108 +245,6 @@ def read_lat_longs(overrides=None, use_defaults=True):
 
 def read_colors(overrides=None, use_defaults=True):
     return ColorParser(mapping_filename=overrides, use_defaults=use_defaults).mapping
-
-def write_VCF_translation(prot_dict, vcf_file_name, ref_file_name):
-    """
-    Writes out a VCF-style file (which seems to be minimally handleable
-    by vcftools and pyvcf) of the AA differences between sequences and the reference.
-    This is a similar format created/used by read_in_vcf except that there is one
-    of these dicts (with sequences, reference, positions) for EACH gene.
-
-    Also writes out a fasta of the reference alignment.
-
-    EBH 12 Dec 2017
-    """
-    import numpy as np
-
-    #for the header
-    seqNames = list(prot_dict[list(prot_dict.keys())[0]]['sequences'].keys())
-
-    #prepare the header of the VCF & write out
-    header=["#CHROM","POS","ID","REF","ALT","QUAL","FILTER","INFO","FORMAT"]+seqNames
-    with open(vcf_file_name, 'w', encoding='utf-8') as the_file:
-        the_file.write( "##fileformat=VCFv4.2\n"+
-                        "##source=NextStrain_Protein_Translation\n"+
-                        "##FORMAT=<ID=GT,Number=1,Type=String,Description=\"Genotype\">\n")
-        the_file.write("\t".join(header)+"\n")
-
-    refWrite = []
-    vcfWrite = []
-
-    #go through for every gene/protein
-    for fname, prot in prot_dict.items():
-        sequences = prot['sequences']
-        ref = prot['reference']
-        positions = prot['positions']
-
-        #write out the reference fasta
-        refWrite.append(">"+fname)
-        refWrite.append(ref)
-
-        #go through every variable position
-        #There are no deletions here, so it's simpler than for VCF nuc sequenes!
-        for pi in positions:
-            pos = pi+1 #change numbering to match VCF not python
-            refb = ref[pi] #reference base at this position
-
-            #try/except is (much) faster than list comprehension!
-            pattern = []
-            for k,v in sequences.items():
-                try:
-                    pattern.append(sequences[k][pi])
-                except KeyError:
-                    pattern.append('.')
-            pattern = np.array(pattern)
-
-            #get the list of ALTs - minus any '.'!
-            uniques = np.unique(pattern)
-            uniques = uniques[np.where(uniques!='.')]
-
-            #Convert bases to the number that matches the ALT
-            j=1
-            for u in uniques:
-                pattern[np.where(pattern==u)[0]] = str(j)
-                j+=1
-            #Now convert these calls to #/# (VCF format)
-            calls = [ j+"/"+j if j!='.' else '.' for j in pattern ]
-            if len(uniques)==0:
-                print("UNEXPECTED ERROR WHILE CONVERTING TO VCF AT POSITION {}".format(str(pi)))
-                break
-
-            #put it all together and write it out
-            output = [fname, str(pos), ".", refb, ",".join(uniques), ".", "PASS", ".", "GT"] + calls
-
-            vcfWrite.append("\t".join(output))
-
-    #write it all out
-    with open(ref_file_name, 'w', encoding='utf-8') as the_file:
-        the_file.write("\n".join(refWrite))
-
-    with open(vcf_file_name, 'a', encoding='utf-8') as the_file:
-        the_file.write("\n".join(vcfWrite))
-
-    if vcf_file_name.lower().endswith('.gz'):
-        import os
-        #must temporarily remove .gz ending, or gzip won't zip it!
-        os.rename(vcf_file_name, vcf_file_name[:-3])
-        call = ["gzip", vcf_file_name[:-3]]
-        run_shell_command(" ".join(call), raise_errors = True)
-
-shquote = shlex.quote
-
-def run_shell_command(cmd, raise_errors=False, extra_env=None):
-    """
-    Run the given command string via Bash with error checking.
-
-    Returns True if the command exits normally.  Returns False if the command
-    exits with failure and "raise_errors" is False (the default).  When
-    "raise_errors" is True, exceptions are rethrown.
-
-    If an *extra_env* mapping is passed, the provided keys and values are
-    overlayed onto the default subprocess environment.
-    """
-    return ShellCommandRunner(cmd, raise_errors=raise_errors, extra_env=extra_env).run()
-
 
 def first_line(text):
     """
